// apps/web/src/lib/easylog/generated-client/models/ProjectBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Easylog No description provided (generated by Openapi Generator
 * https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.2.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator
 * (https://openapi-generator.tech). https://openapi-generator.tech Do not edit
 * the class manually.
 */

import { mapValues } from '../runtime';
/**
 * @export
 * @interface ProjectBody
 */
export interface ProjectBody {
  /**
   * @memberof ProjectBody
   * @type {string}
   */
  name: string;
  /**
   * @memberof ProjectBody
   * @type {string}
   */
  color: string;
  /**
   * @memberof ProjectBody
   * @type {boolean}
   */
  reportVisible: boolean;
  /**
   * @memberof ProjectBody
   * @type {boolean}
   */
  excludeInWorkdays: boolean;
  /**
   * Not required for dateless projects
   *
   * @memberof ProjectBody
   * @type {Date}
   */
  start: Date;
  /**
   * Not required for dateless projects
   *
   * @memberof ProjectBody
   * @type {Date}
   */
  end: Date;
  /**
   * Field in here must exist in `extra_data_fields` on the project datasource
   *
   * @memberof ProjectBody
   * @type {object}
   */
  extraData?: object;
}

/** Check if a given object implements the ProjectBody interface. */
export function instanceOfProjectBody(value: object): value is ProjectBody {
  if (!('name' in value) || value['name'] === undefined) return false;
  if (!('color' in value) || value['color'] === undefined) return false;
  if (!('reportVisible' in value) || value['reportVisible'] === undefined)
    return false;
  if (
    !('excludeInWorkdays' in value) ||
    value['excludeInWorkdays'] === undefined
  )
    return false;
  if (!('start' in value) || value['start'] === undefined) return false;
  if (!('end' in value) || value['end'] === undefined) return false;
  return true;
}

export function ProjectBodyFromJSON(json: any): ProjectBody {
  return ProjectBodyFromJSONTyped(json, false);
}

export function ProjectBodyFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): ProjectBody {
  if (json == null) {
    return json;
  }
  return {
    name: json['name'],
    color: json['color'],
    reportVisible: json['report_visible'],
    excludeInWorkdays: json['exclude_in_workdays'],
    start: new Date(json['start']),
    end: new Date(json['end']),
    extraData: json['extra_data'] == null ? undefined : json['extra_data']
  };
}

export function ProjectBodyToJSON(json: any): ProjectBody {
  return ProjectBodyToJSONTyped(json, false);
}

export function ProjectBodyToJSONTyped(
  value?: ProjectBody | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    name: value['name'],
    color: value['color'],
    report_visible: value['reportVisible'],
    exclude_in_workdays: value['excludeInWorkdays'],
    start: value['start'].toISOString(),
    end: value['end'].toISOString(),
    extra_data: value['extraData']
  };
}